refacto: déplace et simplifie des fixtures utilisées à un seul endroit
